refactor(e2e): slim AC1 helpers facade and section-split province expected lines
Move thin AC1 aliases out of e2e_helpers into topic libraries and split
province panel expected-text builders by section so Slice B of #4075 stays
under physical-line hygiene without changing public entrypoints.
